# src/main/resources/postman/postman.collection.concordia.swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/task": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "operationId": "getAll_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoListTask"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "task-controller"
        ],
        "operationId": "update_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoTask"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "task-controller"
        ],
        "operationId": "insert_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoTask"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/task/{id}": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "operationId": "getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoTask"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "task-controller"
        ],
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoTask"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/title={title}": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "operationId": "getByTitle",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoListTask"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/status={status}": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "operationId": "getByStatus",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "TO_DO",
                "IN_PROGRESS",
                "COMPLETED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoListTask"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/priority={priority}": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "operationId": "getByPriority",
        "parameters": [
          {
            "name": "priority",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "HIGH",
                "EXPIRING",
                "MEDIUM",
                "LOW",
                "DONE",
                "ARCHIVED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoListTask"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/deadLine={deadLine}": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "operationId": "getByDeadLine",
        "parameters": [
          {
            "name": "deadLine",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoListTask"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "TaskPostRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "HIGH",
              "EXPIRING",
              "MEDIUM",
              "LOW",
              "DONE",
              "ARCHIVED"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "TO_DO",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "deadLine": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResponseDtoTask": {
        "type": "object",
        "properties": {
          "httpRequest": {
            "type": "string",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "PATCH",
              "DELETE",
              "COPY",
              "HEAD",
              "OPTIONS",
              "LINK",
              "UNLINK",
              "PURGE",
              "LOCK",
              "UNLOCK",
              "PROPFIND",
              "VIEW"
            ]
          },
          "httpResponse": {
            "type": "string",
            "enum": [
              "CONTINUE",
              "SWITCHING_PROTOCOLS",
              "PROCESSING",
              "EARLY_HINTS",
              "RESPONSE_IS_STALE",
              "REVALIDATION_FAILED",
              "DISCONNECTED_OPERATION",
              "HEURISTIC_EXPIRATION",
              "MISCELLANEOUS_WARNING",
              "OK",
              "AUTHORIZED",
              "CREATED",
              "UPDATED",
              "DELETED",
              "LOADED",
              "ACCEPTED",
              "NON_AUTHORITATIVE_INFORMATION",
              "NO_CONTENT",
              "RESET_CONTENT",
              "PARTIAL_CONTENT",
              "MULTI_STATUS",
              "ALREADY_REPORTED",
              "TRANSFORMATION_APPLIED",
              "IM_USED",
              "MISCELLANEOUS_PERSISTENT_WARNING",
              "MULTIPLE_CHOICES",
              "MOVED_PERMANENTLY",
              "FOUND",
              "SEE_OTHER",
              "NOT_MODIFIED",
              "USE_PROXY",
              "SWITCH_PROXY",
              "TEMPORARY_REDIRECT",
              "PERMANENT_REDIRECT",
              "FAILED",
              "BAD_REQUEST",
              "UNAUTHORIZED",
              "PAYMENT_REQUIRED",
              "FORBIDDEN",
              "NOT_CREATED",
              "NOT_UPDATED",
              "NOT_DELETED",
              "NOT_LOADED",
              "NOT_ACCEPTED",
              "NOT_FOUND",
              "SITE_NOT_FOUND",
              "ISAPI_OR_CGI_RESTRICTION",
              "MIME_TYPE_RESTRICTION",
              "NO_HANDLER_CONFIGURED",
              "DENIED_BY_REQUEST_FILTERING_CONFIGUARTION",
              "VERB_DENIED",
              "FILE_EXTENSION_DENIED",
              "HIDDEN_NAMESPACE",
              "FILE_ATTRIBUTE_HIDDEN",
              "REQUEST_HEADER_TOO_LONG",
              "REQUEST_CONTAINS_DOUBLE_ESCAPE_SEQUENCE",
              "REQUEST_CONTAIN_HIGH_BIT_CHARACTERS",
              "CONTENT_LENGTH_TOO_LARGE",
              "REQUEST_URL_TOO_LONG",
              "QUERY_STRING_TOO_LONG",
              "DAV_REQUEST_SENT_TO_THE_STATIC_FILE_HANDLER",
              "DYNAMIC_CONTENT_MAPPED_TO_THE_STATIC_FILE_HANDLER",
              "QUERY_STRING_SEQUENCE_DENIED",
              "DENIED_BY_FILTERING_RULE",
              "TOO_MANY_URL_SEGMENTS",
              "METHOD_NOT_ALLOWED",
              "NOT_ACCEPTABLE",
              "PROXY_AUTHENTICATION_REQUIRED",
              "REQUESTED_TIMEOUT",
              "CONFLICT",
              "GONE",
              "LENGTH_REQUIRED",
              "PRECONDITION_FAILED",
              "PAYLOAD_TOO_LARGE",
              "URI_TOO_LONG",
              "UNSUPPORTED_MEDIA_TYPE",
              "RANGE_NOT_SATISFIABLE",
              "EXPECTATION_FAILED",
              "IM_A_TEAPOT",
              "PAGE_EXPIRED",
              "METHOD_FAILURE",
              "MISSDIRECTED_REQUEST",
              "UNPROCESSABLE_ENTITY",
              "LOCKED",
              "FAILED_DEPENDENCY",
              "TOO_EARLY",
              "UPGRADE_REQUIRED",
              "PRECONDITION_REQUIRED",
              "TOO_MANY_REQUESTS",
              "REQUEST_HEADER_FIELD_TOO_LARGE",
              "LOGIN_TIME_OUT",
              "NO_RESPONSE",
              "RETRY_WITH",
              "BLOCKED_BY_WINDOWS_PARENTAL_CONTROLS",
              "UNAVAILABLE_FOR_LEGAL_REASONS",
              "SSL_CERTIFICATE_ERROR",
              "SSL_CERTIFICATE_REQUIRED",
              "HTTP_REQUEST_SENT_TO_HTTPS_PORT",
              "INVALID_TOKEN",
              "TOKEN_REQUIRED",
              "CLIENT_CLOSED_REQUEST",
              "NOT_CONTINUE",
              "NOT_SWITCHING_PROTOCOLS",
              "NOT_PROCESSING",
              "INTERNAL_SERVER_ERROR",
              "NOT_IMPLEMENTED",
              "BAD_GATEWAY",
              "SERVICE_UNAVAILABLE",
              "GATEWAY_TIMEOUT",
              "HTTP_VERSION_NOT_SUPPORTED",
              "VARIANT_ALSO_NEGOTIATES",
              "INSUFFICIENT_STORAGE",
              "LOOP_DETECTED",
              "BANDWIDTH_LIMIT_EXCEEDED",
              "NOT_EXTENDED",
              "NETWORK_AUTHENTICATION_REQUIRED",
              "WEB_SERVER_RETURNED_AN_UNKNOWN_ERROR",
              "WEB_SERVER_IS_DOWN",
              "CONNECTION_TIMED_OUT",
              "WEB_SERVER_IS_UNREACHABLE",
              "A_TIME_OUT_OCCURRED",
              "SSL_HANDSAHKE_FAILED",
              "INVALID_SSL_CERTIFICATE",
              "RAILGUN_ERROR",
              "SITE_IS_OVERLOADED",
              "SITE_IS_FROZEN",
              "NETWORK_READ_TIME_OUT_ERROR",
              "NETWORK_CONNECT_TIME_OUT_ERROR"
            ]
          },
          "code": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/Task"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "HIGH",
              "EXPIRING",
              "MEDIUM",
              "LOW",
              "DONE",
              "ARCHIVED"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "TO_DO",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "deadLine": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ResponseDtoListTask": {
        "type": "object",
        "properties": {
          "httpRequest": {
            "type": "string",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "PATCH",
              "DELETE",
              "COPY",
              "HEAD",
              "OPTIONS",
              "LINK",
              "UNLINK",
              "PURGE",
              "LOCK",
              "UNLOCK",
              "PROPFIND",
              "VIEW"
            ]
          },
          "httpResponse": {
            "type": "string",
            "enum": [
              "CONTINUE",
              "SWITCHING_PROTOCOLS",
              "PROCESSING",
              "EARLY_HINTS",
              "RESPONSE_IS_STALE",
              "REVALIDATION_FAILED",
              "DISCONNECTED_OPERATION",
              "HEURISTIC_EXPIRATION",
              "MISCELLANEOUS_WARNING",
              "OK",
              "AUTHORIZED",
              "CREATED",
              "UPDATED",
              "DELETED",
              "LOADED",
              "ACCEPTED",
              "NON_AUTHORITATIVE_INFORMATION",
              "NO_CONTENT",
              "RESET_CONTENT",
              "PARTIAL_CONTENT",
              "MULTI_STATUS",
              "ALREADY_REPORTED",
              "TRANSFORMATION_APPLIED",
              "IM_USED",
              "MISCELLANEOUS_PERSISTENT_WARNING",
              "MULTIPLE_CHOICES",
              "MOVED_PERMANENTLY",
              "FOUND",
              "SEE_OTHER",
              "NOT_MODIFIED",
              "USE_PROXY",
              "SWITCH_PROXY",
              "TEMPORARY_REDIRECT",
              "PERMANENT_REDIRECT",
              "FAILED",
              "BAD_REQUEST",
              "UNAUTHORIZED",
              "PAYMENT_REQUIRED",
              "FORBIDDEN",
              "NOT_CREATED",
              "NOT_UPDATED",
              "NOT_DELETED",
              "NOT_LOADED",
              "NOT_ACCEPTED",
              "NOT_FOUND",
              "SITE_NOT_FOUND",
              "ISAPI_OR_CGI_RESTRICTION",
              "MIME_TYPE_RESTRICTION",
              "NO_HANDLER_CONFIGURED",
              "DENIED_BY_REQUEST_FILTERING_CONFIGUARTION",
              "VERB_DENIED",
              "FILE_EXTENSION_DENIED",
              "HIDDEN_NAMESPACE",
              "FILE_ATTRIBUTE_HIDDEN",
              "REQUEST_HEADER_TOO_LONG",
              "REQUEST_CONTAINS_DOUBLE_ESCAPE_SEQUENCE",
              "REQUEST_CONTAIN_HIGH_BIT_CHARACTERS",
              "CONTENT_LENGTH_TOO_LARGE",
              "REQUEST_URL_TOO_LONG",
              "QUERY_STRING_TOO_LONG",
              "DAV_REQUEST_SENT_TO_THE_STATIC_FILE_HANDLER",
              "DYNAMIC_CONTENT_MAPPED_TO_THE_STATIC_FILE_HANDLER",
              "QUERY_STRING_SEQUENCE_DENIED",
              "DENIED_BY_FILTERING_RULE",
              "TOO_MANY_URL_SEGMENTS",
              "METHOD_NOT_ALLOWED",
              "NOT_ACCEPTABLE",
              "PROXY_AUTHENTICATION_REQUIRED",
              "REQUESTED_TIMEOUT",
              "CONFLICT",
              "GONE",
              "LENGTH_REQUIRED",
              "PRECONDITION_FAILED",
              "PAYLOAD_TOO_LARGE",
              "URI_TOO_LONG",
              "UNSUPPORTED_MEDIA_TYPE",
              "RANGE_NOT_SATISFIABLE",
              "EXPECTATION_FAILED",
              "IM_A_TEAPOT",
              "PAGE_EXPIRED",
              "METHOD_FAILURE",
              "MISSDIRECTED_REQUEST",
              "UNPROCESSABLE_ENTITY",
              "LOCKED",
              "FAILED_DEPENDENCY",
              "TOO_EARLY",
              "UPGRADE_REQUIRED",
              "PRECONDITION_REQUIRED",
              "TOO_MANY_REQUESTS",
              "REQUEST_HEADER_FIELD_TOO_LARGE",
              "LOGIN_TIME_OUT",
              "NO_RESPONSE",
              "RETRY_WITH",
              "BLOCKED_BY_WINDOWS_PARENTAL_CONTROLS",
              "UNAVAILABLE_FOR_LEGAL_REASONS",
              "SSL_CERTIFICATE_ERROR",
              "SSL_CERTIFICATE_REQUIRED",
              "HTTP_REQUEST_SENT_TO_HTTPS_PORT",
              "INVALID_TOKEN",
              "TOKEN_REQUIRED",
              "CLIENT_CLOSED_REQUEST",
              "NOT_CONTINUE",
              "NOT_SWITCHING_PROTOCOLS",
              "NOT_PROCESSING",
              "INTERNAL_SERVER_ERROR",
              "NOT_IMPLEMENTED",
              "BAD_GATEWAY",
              "SERVICE_UNAVAILABLE",
              "GATEWAY_TIMEOUT",
              "HTTP_VERSION_NOT_SUPPORTED",
              "VARIANT_ALSO_NEGOTIATES",
              "INSUFFICIENT_STORAGE",
              "LOOP_DETECTED",
              "BANDWIDTH_LIMIT_EXCEEDED",
              "NOT_EXTENDED",
              "NETWORK_AUTHENTICATION_REQUIRED",
              "WEB_SERVER_RETURNED_AN_UNKNOWN_ERROR",
              "WEB_SERVER_IS_DOWN",
              "CONNECTION_TIMED_OUT",
              "WEB_SERVER_IS_UNREACHABLE",
              "A_TIME_OUT_OCCURRED",
              "SSL_HANDSAHKE_FAILED",
              "INVALID_SSL_CERTIFICATE",
              "RAILGUN_ERROR",
              "SITE_IS_OVERLOADED",
              "SITE_IS_FROZEN",
              "NETWORK_READ_TIME_OUT_ERROR",
              "NETWORK_CONNECT_TIME_OUT_ERROR"
            ]
          },
          "code": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "body": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          }
        }
      }
    }
  }
}